+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix PHP 8.4 deprecation warnings #534 @[robertoschwald](https://github.com/robertoschwald)
 - date formatting on all rows
 - Exclude hidden columns from time aggregation options
+- Correct visible dimension mapping for time aggregation
 
 ## 5.8.0 - 2025-07-29
 ### Added